+++ xskillscore/core/resampling.py
@@ -222,6 +222,13 @@ def resample_iterations_idx(
     transpose_kwargs = (
         {"transpose_coords": False} if isinstance(forecast, xr.DataArray) else {}
     )
+    # bug fix if singleton dimension
+    singleton_dims = []
+    for d in forecast.dims:
+        if forecast.sizes[d] == 1:
+            singleton_dims.append(d)
+            forecast = forecast.isel({d: [0] * 2})
+
     # bug fix when only one iteration
     if iterations == 1:
         forecast_smp = forecast.isel({dim: idx_da.isel(iteration=0, drop=True).values})
@@ -238,4 +245,6 @@ def resample_iterations_idx(
         forecast_smp = forecast_smp.isel({dim: slice(None, dim_max)})
     if dim_coord_set:
         del forecast_smp.coords[dim]
+    for d in singleton_dims:
+        forecast_smp = forecast_smp.isel({d: [0]})
     return forecast_smp